requirements.txt
streamlit
torch
transformers
tensorflow
pillow
numpy
safetensors
gdown